+++ client/plots/sampleScatter.js
@@ -2,22 +2,6 @@ import { getCompInit, copyMerge } from '../rx'
 import { fillTermWrapper } from '#termsetting'
 import { Menu } from '#dom/menu'
 import { controlsInit } from './controls'
-import {
-	symbol,
-	symbolCircle,
-	symbolTriangle,
-	symbolCross,
-	symbolSquare,
-	symbolWye,
-	symbolDiamond,
-	symbolDiamond2,
-	symbolSquare2,
-	symbolTriangle2,
-	symbolStar,
-	symbolPlus,
-	symbolAsterisk,
-	symbolX
-} from 'd3-shape'
 import { setRenderers } from './sampleScatter.renderer'
 import { setInteractivity } from './sampleScatter.interactivity'
 import { getActiveCohortStr } from '../mass/charts'
@@ -43,26 +27,32 @@ NOTE
 */
 export const minShapeSize = 0.2
 export const maxShapeSize = 4
+
+export const shapes = [
+	'M 8,8 m 8,0 a 8,8 0 1,0 -16,0 a 8,8 0 1,0 16,0', //circle filled
+	'M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14m0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16', //circle empty
+
+	'M7.022 1.566a1.13 1.13 0 0 1 1.96 0l6.857 11.667c.457.778-.092 1.767-.98 1.767H1.144c-.889 0-1.437-.99-.98-1.767z', //triangle filled
+	'M7.938 2.016A.13.13 0 0 1 8.002 2a.13.13 0 0 1 .063.016.15.15 0 0 1 .054.057l6.857 11.667c.036.06.035.124.002.183a.2.2 0 0 1-.054.06.1.1 0 0 1-.066.017H1.146a.1.1 0 0 1-.066-.017.2.2 0 0 1-.054-.06.18.18 0 0 1 .002-.183L7.884 2.073a.15.15 0 0 1 .054-.057m1.044-.45a1.13 1.13 0 0 0-1.96 0L.165 13.233c-.457.778.091 1.767.98 1.767h13.713c.889 0 1.438-.99.98-1.767z', //triangle empty
+
+	'M6.95.435c.58-.58 1.52-.58 2.1 0l6.515 6.516c.58.58.58 1.519 0 2.098L9.05 15.565c-.58.58-1.519.58-2.098 0L.435 9.05a1.48 1.48 0 0 1 0-2.098z', //diamond filled
+	'M14 1a1 1 0 0 1 1 1v12a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1zM2 0a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2z', //diamond empty
+
+	'M0 2a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v12a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2z', //square filled
+	'M6.95.435c.58-.58 1.52-.58 2.1 0l6.515 6.516c.58.58.58 1.519 0 2.098L9.05 15.565c-.58.58-1.519.58-2.098 0L.435 9.05a1.48 1.48 0 0 1 0-2.098zm1.4.7a.495.495 0 0 0-.7 0L1.134 7.65a.495.495 0 0 0 0 .7l6.516 6.516a.495.495 0 0 0 .7 0l6.516-6.516a.495.495 0 0 0 0-.7L8.35 1.134z', //square empty
+
+	'M7.685.256a.5.5 0 0 1 .63 0l7.421 6.03a.5.5 0 0 1 .162.538l-2.788 8.827a.5.5 0 0 1-.476.349H3.366a.5.5 0 0 1-.476-.35L.102 6.825a.5.5 0 0 1 .162-.538l7.42-6.03Z', //pentagon filled
+	'M7.685 1.545a.5.5 0 0 1 .63 0l6.263 5.088a.5.5 0 0 1 .161.539l-2.362 7.479a.5.5 0 0 1-.476.349H4.099a.5.5 0 0 1-.476-.35L1.26 7.173a.5.5 0 0 1 .161-.54l6.263-5.087Zm8.213 5.28a.5.5 0 0 0-.162-.54L8.316.257a.5.5 0 0 0-.631 0L.264 6.286a.5.5 0 0 0-.162.538l2.788 8.827a.5.5 0 0 0 .476.349h9.268a.5.5 0 0 0 .476-.35l2.788-8.826Z', //pentagon empty
+
+	'M8 2a.5.5 0 0 1 .5.5v5h5a.5.5 0 0 1 0 1h-5v5a.5.5 0 0 1-1 0v-5h-5a.5.5 0 0 1 0-1h5v-5A.5.5 0 0 1 8 2', //plus
+	'M14 10a6 6 0 0 1-12 0C2 5.686 5 0 8 0s6 5.686 6 10', //egg filled
+	'M8 15a5 5 0 0 1-5-5c0-1.956.69-4.286 1.742-6.12.524-.913 1.112-1.658 1.704-2.164C7.044 1.206 7.572 1 8 1s.956.206 1.554.716c.592.506 1.18 1.251 1.704 2.164C12.31 5.714 13 8.044 13 10a5 5 0 0 1-5 5m0 1a6 6 0 0 0 6-6c0-4.314-3-10-6-10S2 5.686 2 10a6 6 0 0 0 6 6' //egg empty
+]
+
 class Scatter {
 	constructor() {
 		this.type = 'sampleScatter'
 		this.lassoOn = false
-		const mySymbols = [
-			symbolCircle,
-			symbolSquare,
-			symbolCross,
-			symbolWye,
-			symbolTriangle,
-			//symbolDiamond,
-			symbolDiamond2,
-			symbolStar,
-			symbolPlus,
-			symbolSquare2,
-			symbolTriangle2,
-			symbolX,
-			symbolAsterisk
-		]
-		this.symbols = mySymbols.map(s => symbol(s))
 		this.zoom = 1
 		this.startGradient = {}
 		this.stopGradient = {}
@@ -270,7 +260,7 @@ class Scatter {
 			type: 'number',
 			chartType: 'sampleScatter',
 			settingsKey: 'size',
-			title: 'Sample scale, increase to make the samples larger',
+			title: 'Sample size, represents the factor used to scale the sample',
 			min: 0
 		}
 		const minShapeSizeOption = {
@@ -335,7 +325,7 @@ class Scatter {
 				type: 'number',
 				chartType: 'sampleScatter',
 				settingsKey: 'opacity',
-				title: 'It represents the opacity of the symbols',
+				title: 'It represents the opacity of the elements',
 				min: 0,
 				max: 1,
 				step: 0.1
